Re-design for ShardingSphere groovy expr
